requirements.txt
Django==3.1.3
pytz==2019.2
social-auth-app-django==3.1.0
sqlparse==0.3.0
django-markdownx==3.0.1
django-environ==0.4.5
psycopg2-binary==2.8.6
six==1.15.0
Pillow==8.0.1
redis==3.5.3
